Split the `BackendMenuListener` class into two separate classes
